various stackmaptable bug fixes
StackMapTable documentation in oracles specification is a mess and
partly wrong, so major fixes had to be applied.

+++ src/com/mitp0sh/jaclaff/attributes/Attributes.java
@@ -19,7 +19,7 @@ public class Attributes
 	
 	public int getAttributesCount()
 	{
-		return (short)attributes.length;
+		return attributes.length;
 	}
 	
 	public AbstractAttribute[] getAttributes()
